+++ data/tasks.txt
@@ -1,0 +1,6 @@
+1. [T] [ ] task
+2. [T] [ ] tak3
+3. [T] [ ] tisk
+4. [E] [ ] dance
+5. [E] [ ] dance
+6. [E] [ ] dance
