+++ src/main/java/duke/Duke.java
@@ -1,0 +1,99 @@
+package duke;
+
+import duke.exceptions.DukeExceptions;
+
+import java.io.IOException;
+import java.util.Scanner;
+import javafx.util.Pair;
+
+
+/**
+ * Programme which serves as an interactive checklist.
+ */
+public class Duke {
+
+    private duke.Storage storage = new duke.Storage();
+    private duke.TaskList tasks;
+    private duke.Ui ui;
+    private duke.Parser parser = new duke.Parser();
+
+    /**
+     * Initializes Duke.
+     * @param filePath represents the path of the file and existing tasks to be loaded
+     *                 if already present.
+     */
+    public Duke(String filePath) {
+        ui = new duke.Ui();
+        storage = new duke.Storage(filePath);
+        try {
+            tasks = storage.load();
+            //tasks = new duke.TaskList();
+        } catch (Exception e) {
+            ui.showLoadingError();
+            tasks = new duke.TaskList();
+        }
+    }
+
+
+    /**
+     * Initializes Duke.
+     *
+     */
+    public Duke() {
+        ui = new duke.Ui();
+        try {
+            tasks = storage.load();
+            //tasks = new duke.TaskList();
+        } catch (Exception e) {
+            ui.showLoadingError();
+            tasks = new duke.TaskList();
+        }
+    }
+
+    /**
+     * Runs through live user input to add, edit the tasks.
+     */
+    void run() throws DukeExceptions, IOException {
+        ui.start();
+
+        Scanner sc = new Scanner(System.in);
+
+        while (sc.hasNextLine()) {
+            String values = sc.nextLine();
+            Pair<duke.TaskList, String> pair = parser.parse(ui, tasks, values);
+            tasks = pair.getKey();
+            String output = pair.getValue();
+            System.out.println(output);
+            if (values.equals("bye")) {
+                return;
+            }
+            storage.save(tasks);
+        }
+    }
+
+    public String getResponse(String input) throws IOException, DukeExceptions {
+
+        try {
+            Pair<duke.TaskList , String> pair = parser.parse(ui, tasks, input);
+            tasks = pair.getKey();
+            String output = pair.getValue();
+            System.out.println(output);
+            if (input.equals("bye")) {
+                return "BYEBYE PIKACHU WILL MISS U";
+            }
+            storage.save(tasks);
+            return "Pikachu heard you say: "+input+ System.lineSeparator() + output;
+        } catch(DukeExceptions e) {
+            return e.getMessage();
+        }
+
+    }
+}
+
+
+
+
+
+
+
+
