+++ src/main/java/duke/exceptions/DukeExceptions.java
@@ -1,0 +1,9 @@
+package duke.exceptions;
+
+public class DukeExceptions extends Exception {
+    public DukeExceptions(String output) {
+            super(output);
+    }
+}
+
+
