+++ src/main/java/duke/exceptions/DukeInvalidInput.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public class DukeInvalidInput extends Exception{
+    public DukeInvalidInput() {
+        super("ENTER A VALID INPUT! ^_^");
+    }
+}
